+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Snapshot tests for `BuyButton`.
 - Snapshot tests for `AvailabilitySubscriber`.
 - Snapshot tests for `Animation`.
+- Snapshot tests for `CollectionBadges`.
 
 ## [3.17.2] - 2019-02-18